Add support for indexing arrays
Summary:
This required a rather large overhaul to the way I'm handling lvalues.
Mostly this was just because I didn't realize it's not allowed to push to the
stack with a live base pointer, so i needed a way of allowing the member
instructions to happen all at once, once the rest of the stack is prepped. This
is why the class `MinstrSeq` appears

This pulls the lvalue code into its own module, and adds the notion of a
BaseValue (which is just something we can take a base pointer to)--since array
cells are not lvalues, but you can take a base pointer to them.

I also re-evaluated the zend suite and quite a few more pass now that we have
array support. Note that the list wasn't sorted before--this is actually an
all-green diff for that file after reordering

Reviewed By: swtaarrs

Differential Revision: D5498319

fbshipit-source-id: 7a7bddc672cd9abd453a2cb30567d0fac1bdb43b