Make compilation stateless: the Great Refactor
Summary:
Trying to implement `finally` has been an enormous struggle--especially since
the current architecture keeps a lot of compilation state in mutable positions
on the compiler object: this means that trying to juggle multiple states (e.g.
to add code to a catch region but also to another region, all while keeping
track of which loops, try/finally blocks are in scope, etc. ) gets really
tricky and bug-prone

This adds a notion of a CFG which is essentially a composable chunk of bytecode
that we can return directly from compilation methods. Add a little jiggering
and we can make the whole compilation process stateless :)

There's two regressions here:

 - The resulting assembly has way more labels than it needs; this is easily
   fixed and I'll put up a diff shortly to merge blocks or at least remove redundant labels
 - I don't track local names anymore--this needs to be fixed since I'll need
   unnamed locals to implement try/finally, but this should be done by walking
   the CFG instead of manually tracking the set of used local names

refactor_everything

Reviewed By: swtaarrs, mofarrell

Differential Revision: D5551832

fbshipit-source-id: f16e1170c1435b4ba7864d6c3d053f2494098c28