Treat Shapes.hhi specially: It should always promote shapes to use unknown fields
Summary:
We want to treat Shapes.hhi specially, by having its arguments accept shapes with unknown fields, even though the source code for Shapes.hhi does not indicate as such. This is to preserve compatibility with the open source Hack, which has no knowledge of the unknown fields syntax.

It looks like we do standard type checking of the parameters when a function call is found, which happens even before any code reaches typing_shapes.ml.

Consequently, we still need to update the typechecker code to treat Shapes.hhi specially.

This solution accomplishes that, by promoting shapes in Shapes.hhi to be as if they supported unknown fields.

Reviewed By: kmeht

Differential Revision: D5589850

fbshipit-source-id: 971ba01164172d0d35db006d6483474b734fb26b