Get rid of O(n^2) behavior in Option::FilterInputs
Summary:
I noticed that finding the list of files to parse was taking twice as
long as it took to actually parse them. Most of that time was spent in
Option::Filter. Part of that is spent in preg_match, but a big chunk
was spent just removing elements from a vector one at a time - each
removal being O(n).

Reviewed By: ricklavoie, alexeyt

Differential Revision: D5598986

fbshipit-source-id: 62b869d41cbe9205c4247ee4ad1efac906b0bc26