Simplify lambda analysis
Summary:
resolve_lambda_names was doing a full pass over the entire ast just to
collect an "implicit this" flag for lambdas, and then update the
lambda's capture variables.

We can collect "implicit this" as part of the regular analysis
pass. The only complication is that we have to wait until the function
containing the lambda has been fully processed, in case any implicitly
captured variables haven't been seen at the point the lambda is
invoked. Generally, this is an error condition anyway, but if the
lambda is in a loop, it could matter. To solve that, this diff adds a
CompactVector of lambdas to each file scope, which gets processed (and
emptied) when the whole file scope has been analyzed.

resolve_lambda_names also got scoping wrong, so eg given

  class X {
    function foo($a, $b) {
      return () ==> {
        return function () {
          return () ==> {
            $a = 42;
            $b = 100;
            ...
          }
        };
      };
    }
  }

It would walk upwards through the regular closure to foo, and assume
that $a and $b were captured - pointlessly adding use vars to the
lambda, and otherwise-unused locals to the closure. I fixed that so it
only walks through nested lambdas.

Note also that resolve_lambda_names was called both at analyze all
time, and at analyze final time, even though it never actually updated
anything in the second call.

Reviewed By: ricklavoie

Differential Revision: D5590272

fbshipit-source-id: 54da62ba0909a7fb090bb8f4ff4a0e1c3bad32d3