disallow jump to first bytecode in function
Summary:
Verifier now disallows jumping to the first bytecode in a function unless the origin of the jump is in a DV initializer. This is to prevent the JIT from re-running startup bytecode again associated with the first bytecode in the function.

This also makes a small change to hhbbc to make it compatible with this verification. If an entry block in a function is just a Nop, this Nop is replaced with an EntryNop. This way the block won't be removed by the emitter, and thus jumps to the second block will not be considered jumps to the first bytecode.

Reviewed By: paulbiss

Differential Revision: D5584122

fbshipit-source-id: a802226da37b0496d8ed7d3329c448a8b73f8f22