Create libhackfmt module
Summary:
Previously, a lot of application logic lived in hackfmt.ml, which was getting to be a bit unwieldy. Additionally, consumers of hackfmt needed to use functions in the Hack_format, Chunk_builder, and Line_splitter modules, since there was no public API chaining these implementation details together outside of hackfmt.ml, which is not a library.

This change moves all this application logic into the hackfmt ocaml_library, in a new module named Libhackfmt. This allows the CLI tool (hackfmt.ml) to only be concerned with consuming CLI arguments, performing IO, and logging to Scuba, and makes it easier to use hackfmt from other OCaml applications.

In order to avoid a circular dependency between hackfmt and line_splitter, it was necessary to remove the line_splitter ocaml_library and move its modules to the hackfmt library. I don't think this does any harm--as far as I know, the line_splitter library was just mimicking the filesystem structure of dartfmt, and did not exist because of a technical requirement.

Reviewed By: arxanas

Differential Revision: D5529205

fbshipit-source-id: 9bc3cb70c43c047ab94f81990b7542f4c2c4872d