Never add trailing commas in PHP files
Summary: PHP's parser does not allow trailing commas in function definitions or invocations, so this change prevents us from adding them. In order to do so, this diff introduces a Format_env record to carry configuration information through formatting functions. In the future, this record will be extended with other options (such as indentation width and whether to indent with tabs or spaces).

Reviewed By: arxanas

Differential Revision: D5584202

fbshipit-source-id: 2738265163b206579b68a0886719f97eb0fc40e0