Use independent split for CaseLabel
Summary:
Previously, we were using a plain Split, which is governed by the nearest containing WithRule block. In the added test case, the nearest containing WithRule is the one governing the splits around the lambda expression. Since those splits are broken on, line breaks were added between the `case` keyword and label for all CaseLabels (instead of only the one that exceeds the line length limit). This change ensures that this split is only broken on when it helps reduce overflow, instead of erroneously breaking because of the behavior of some irrelevant parent rule.

Depends on D5567179

Differential Revision: D5567286

fbshipit-source-id: ff342ba3dd1c908012cabe98cd5149d8879c43f5