Rename Fmt_node to Doc
Summary: Fmt_node was always a poor name. I wasn't creative enough at the time to come up with something better. Philip Wadler's pretty printer uses the term Doc for an intermediate representation similar to Fmt_node, and I think it's a fitting name--it's a tree describing a document with rules about how it should be formatted.

Differential Revision: D5567500

fbshipit-source-id: 0f22e5b0077ec39ea258008c81cf742da44da130