Fix indentation of comments at the end of a SwitchSection
Summary:
It's conventional to indent FALLTHROUGH comments to the same level as the statements of a SwitchSection, but since they are part of the leading trivia of the following CaseLabel, they were previously indented to the same level as the CaseLabel.

This change removes all trivia up to the first FallThrough trivia from the first CaseLabel in a SwitchSection and indents it, causing the FALLTHROUGH comment (and all comments preceding it) to be indented to the same level as the statements above them.

Depends on D5575486

Reviewed By: arxanas

Differential Revision: D5607144

fbshipit-source-id: 58109f7560661faf900124ef94eaf4e1ba203565