Find files to parse in parallel
Summary:
Finding the files to compile can be quite expensive, especially if
there are many PackageExcludePatterns. This converts the existing
parse queue to include finding the files. The workitems can now be
files or directories. Files are parsed as before, directories are scanned
(non-recursively), and new workitems (again, directories or files) are
created from their contents.

Reviewed By: alexeyt

Differential Revision: D5606603

fbshipit-source-id: 36903cb9be877a8cb3a61b14b603c5d3b96d823f