Global Name Autocomplete
Summary:
This diff is part of the FFP powered autocomplete project.

To achieve parity with the existing autocomplete, we wish to complete global names, which includes namespaces, functions, and classes defined in other files.

The Hack server regularly indexes global functions/classes within the project into an AutocompleteTrie, which is accessible through the HackSearch Service. We query the HackSearchService directly to find what names are available globally. Currently, we only grab the functions and classes available. Namespaces will be done at a future date.

Classes, functions, interfaces, and traits are all valid anywhere an expression is valid since classes, interfaces, and traits can contain constants or static properties or functions that are expressions.

Interfaces and classes can both be used as a type annotation, while a trait cannot.

After the keyword "new", only a class is valid since traits and interfaces cannot be instantiated.

After the keyword "extends", we expect classes, and after the keyword "implements", we expect interfaces. After the keyword "use" inside a class body, we expect traits.

This appears to work when manually testing, but in hh_single_type_check, which is what the tests run, the AutocompleteTrie doesn't get updated so the global list of names is empty and it's hard to write unit tests for this.

A future TODO is to get the names of classes/functions in the current file and complete those to achieve parity with the existing system. Once we do that, we will be able to write some unit tests for this.

Reviewed By: arxanas

Differential Revision: D5566135

fbshipit-source-id: fb082392248e434c5c553d9d1574872b6dc2cb3e