Do completions when directly in front of a close paren
Summary:
The goal of this diff is to autocomplete situations like these:

```
class C {
  public function foo(s|)
}
```

```
function bar(){
  if(|)
}
```

The tricky part of this is that the marked locations are associated with the right paren in the syntax tree. The algorithm for how this is handled is as follows:
- If the location is identified as immediately before the text of the token (when the cursor is at the start_offset of the token), then we check if the token is a symbol token.
- If the token is not a symbol, then we continue as before. If it is a symbol, then we consider the cursor to be in the leading trivia of the node and we continue on the existing code paths that handle leading trivia.
- Furthermore, if we discover that we are in front of a close paren, then we need to determine whether or not to use the preceding token to filter our results. To do this, we use a regex for valid php identifiers taken from here: http://php.net/manual/en/language.variables.basics.php

There is also the edge case of differentiating between the following three situations:
```
1.
if($ |)
2.
if($| )
3.
if($|)
```
This is handled by checking if the predecessor token has trailing trivia. In situation 1, the dollar sign has a trailing trivia of width 1, so we do not do completion here. Situation 2 was already handled by the previous code. Situation 3 sees a previous node with a trailing width of 0, so we do the completion as if we are in the leading trivia of the right paren and complete local variables at this location.

Reviewed By: arxanas

Differential Revision: D5576058

fbshipit-source-id: 56ab09a7a2ee2f616b5c05860c85061f5bde59ac