Fix autocomplete at end of file bug
Summary:
When autocompleting at the end of a file, we move the position back by one so that we end up inside the syntax tree. This can cause issues if moving back by one space causes us to end up inside another token.

This is solved by using the original offset to classify the location of the cursor. However, this involves special cases where the offset is equal to the start offset of the end of file character. If there are extra newlines at the end of the file, there will be an EndOfFile token in the syntax tree which will be outside of the context where we actually wish to complete. To solve this, if the cursor is inside the end of file token, then we consider that location as the leading trivia of the end of file token, leading to the correct expected behavior.

Prior to this, "class" would have been completed in the following example:
```
<?hh
abstract class |
```
This is due to the cursor moving back by one character, then reading "class" as the cursor location with "abstract" as the predecessor. This diff fixes this so we complete nothing in this situation as expected.

Reviewed By: arxanas

Differential Revision: D5577204

fbshipit-source-id: ab44954795409375ce74cd344910562af3396822