Grab classes/functions defined in the same file and use them for autocomplete
Summary:
We do a fold over the syntax tree to gather all of the classes, traits, interfaces, and functions defined in the same file.

This matches the behavior of the current autocomplete. It may seem strange to complete the class we are defining after extends, but this is how the current autocomplete works. It might make sense to look at the parentage of the cursor location and remove the class being defined. i.e.
Don't suggest ClassA in the following example.
```
ClassB {}
ClassA extends Cl|
```

This also allows us to have unit tests for completing classes, traits and interfaces now.

Reviewed By: arxanas

Differential Revision: D5588655

fbshipit-source-id: 6853e5d4ca91af59f7fc0be6197cdff321147032