Use the typechecker to get type information for local variables.
Summary:
Since we want to display type information about the local variables we are completing, we call into the existing autocomplete.

Although this means that the typechecker is run from two different places in the code, the two situations should be mutually exclusive so the typechecker should be run at most once per request. The class based completions run exactly when the preceding token is "::" or "->", and local variable autocomplete happens when inside whitespace or the token to be completed starts with "$".

Potential improvements on this would be to run the typechecker directly and grab the types of variables from the typing_env.

Reviewed By: arxanas

Differential Revision: D5595774

fbshipit-source-id: bfd037080df5a136660069ffee8a9db3aabd37bb