Optimize lookup_param_prep
Summary:
I half fixed this a while ago; if no method takes the param by ref, we
already take a fast path, but if at least one takes it by ref, we end
up going down the slow path, even though we're guaranteed to return
unknown. We only need to take the slow path for parameters beyond the
first 64, which we don't bother to memoize.

Reviewed By: ricklavoie

Differential Revision: D5619473

fbshipit-source-id: c97fc181f870c2efc359f4c378b3945088b1fa29