Kill some scope dependencies
Summary:
hphpc maintains a dependency graph between scopes. Each dependency has
a mask of use-kinds which reflect the kind of dependency. Most of
these were for type-inference, or post-type-inference optimizations,
so we can drop them.

UseKindCaller is a slightly odd case, however. It was used to indicate
that not only should callers be re-evaluated, but that the scope
itself should be too. So although we can drop all explicit
UseKindCaller dependencies, we still need to keep track of when
"UseKindCaller" changes are made to the scope, so that we'll iterate
on it. Since the flag no longer has anything to do with calls, rename
it to UseKindSelf.

Reviewed By: ricklavoie

Differential Revision: D5616105

fbshipit-source-id: b8a5107494dde3e56a498446246b364fb45d3a1d