Allow missing type in catch clause (PHP support)
Summary:
PHP allows for untyped catch clauses, i.e.
```
try {
  throw new Exception();
} catch (/* nothing here */ $_) {
}
```

Previously, even in non-strict modes, the full fidelity parser would consume `$_` as the type and produce an error. We still produce the error, but now we rewind the lexer, to allow the type to be missing. Errors are dealt with later (or not, because PHP).

Reviewed By: andrewjkennedy

Differential Revision: D5592283

fbshipit-source-id: 85591e22a7210a8e469637de1e51bfb343194b65