Avoid Building `anyByRef` in HHBBC Index in Non-repo Mode
Summary:
`method_ref_params_by_name` is a global bit map used for checking if there is any function that uses its n-th function by-ref. This is used in resolving param prep kind.

However, in non-repo mode, without global knowledge, this bitmap is of no use. We avoid building it in non-repo mode. The resolution of prep kind is fixed in D5452355.

Reviewed By: markw65

Differential Revision: D5611635

fbshipit-source-id: b8818fcd1b59d8e36e542c7d85fd4d0e1cde9007