Preserve linebreaks for shape expressions as well as shape types
Summary:
Before, we were preserving linebreaks in shape type literals, such as this:

```
type Foo = shape(
  'foo' => int,
);
```

But we were not doing so for shape expressions, such as this:

```
takes_a_foo(shape(
  'foo' => 1,
));
```

This commit adds support for that.

This commit also renames the test case files from `collection_literal_newlines.php` to `container_literal_newlines.php`, since that's what the function is called in the source code.

Differential Revision: D5624998

fbshipit-source-id: ba21437ca9907b8589534504b85a5c1c1173cb74