Add support for finally
Summary:
What it says on the tin. This is pretty hairy because PHP is really
flexible with what it allows you to do in a try block *and* in a finally block.

To be general, we have to pessimize things a bit and replace exits from the try
block with jumps to the finally handler, and then a switch to return control to
where it should be. An alternative approach would be to just duplicate the
finally code at every exit. This isn't great but it's largely the same approach
as the current emitter built into HHVM.

We could combine this approach with something to optimize the simple cases (no
returns from finally, or no return/goto/breka/continue in try) to get rid of
some of this pessimization

Reviewed By: swtaarrs

Differential Revision: D5575402

fbshipit-source-id: a5c163110c87e6bd4b4e0170fb4c7e7b867660e7