Attempt to fix flaky timeout_test
Summary:
[hphp/hack/test/unit/utils:timeout](https://our.intern.facebook.com/intern/testinfra/testdetail/102431738)
is pretty flaky. Looking at failing test runs, the failing tests tend to look
like this:

1. Spawn another process that is supposed to timeout after X seconds
2. Wait Y seconds (where Y > X)
3. Fail when child process hasn't exited yet.

I'm guessing that these spurious failures are due to spawning a child process
taking surprisingly long when the machine is under stress (like running all the
tests).

We probably could just delete these tests. It's probably sufficient to test
that timeouts work in the current process, without testing that they work in
spawned processes too. However, I'm first going to try and fix this by making
the parent process wait for the child process to start before it starts the
timer.

Yes, this does kind of bias things the other direction - that is now the parent
might start its timer too late instead of too early if the host is slow to
resume the parent process. So it's possible this test will start being flaky in
the other direction (parent process timers failing to fire in time). But let's
try this and see how it does.

Reviewed By: dabek

Differential Revision: D5602215

fbshipit-source-id: 5c56b4d63f26b8b969297c876efd2919eeed0dba