Coroutines should not re-enter themselves before finishing
Summary:
This handles the case where a coroutine will re-enter itself if it calls another coroutine and that other coroutine completes synchronously.

The solution is to not directly call `resume` and `resumeWithException` from `ClosureBase` (only a developer should call those methods). Where we're currently calling those, we actually need to get our hands on a `ActualCoroutineResult` or `SuspendedCoroutineResult`. The caller should propagate the suspension or continue as appropriate.

Instead, we should call `doResume` directly from the lowered code, instead of calling resume or resumeWithException. We actually want the contract exposed by `doResume` instead of the contract exposed by the other two.

Reviewed By: ericlippert, vladima

Differential Revision: D5650371

fbshipit-source-id: b80577426c8dd109e20c6d425313c7b0285beed7