Exactly scan Iterator variables
Summary:
hhbc iterator variables are a union of ArrayIter, MArrayIter,
and CufIter. They have incompatible layouts (pointers at different
locations), and we don't have a fast way to determine whether
an iterator is live or not, much less which kind it is. Hence
they've been conservatively scanned.

To allow exact scanning, we need to know where the pointers are;
it's okay if they're not live. To that end, rearrange fields so
all the pointers overlap with each other, then introduce a
dummy Iter type to allow those pointers to be scanned.

Differential Revision: D5621631

fbshipit-source-id: 022f363f98edf777842aa4202a0ce5ec4a47fe8f