Adding rudimentary misspelling detection to FFP error recovery
Summary:
In this diff, I make a first attempt at adding misspelling detection to syntax errors thrown by the FFP.

However, this is very much a "first attempt:" I'm reasonably happy with the functions I've added to actually do the misspelling detection, but I don't think I've deployed them in the most #impact place. Meaning: I initially thought require_token would be a good place to do misspelling-detection, because it would immediately take effect on a large set of tokens. However...of the 59 tokens that are sometimes 'required,' fewer than five are two or more characters long.

And the five that are are pretty specialty, like: the 'while' of a do/while, or LambdaArrow and ColonColon in very specific contexts. Moreover, due to quirks of the parser structure, the misspelling-detection-in-the-require_[TokenKind] method doesn't actually catch most of these, because in cases like LambdaArrow and ColonColon, the require_[TokenKind] method is only called after the parser has already peeked ahead and verified that those tokens are in fact present.

So, in summary...the way that misspelling detection is deployed here will basically only catch misspellings of the 'while' of a do/while. I'd prefer to leave it in the require_token method, on the off chance that family of methods gets extended to include a wider variety of longer TokenKinds, but I'd also be open to disabling it there and simply deploying the functionality more thoughtfully in a later sequence of diffs.

My goal is to add subsequent diffs to make more effective use of these functions. For example, it would make sense to do misspelling detection over the large, multi-case match statements that sometimes show up in the parsers, like the one in parse_statement in the statement parser.

Differential Revision: D5598818

fbshipit-source-id: 1f9e3aca1d0f5f72a9fb987992d921390f2b7e14