Exploratory implementation of spellcheck in another case
Summary:
The story behind this diff is: when I started it, I thought I saw a clear path to implementing spellcheck in a pretty quick, #impact way. I turned out to be wrong about that. However, this diff does still contain work I had to do to get used to the scope of the problem, and it indeed makes an incremental improvement to parser error recovery. It might be worth accepting just for the latter, and hopefully the former might be useful to someone else down the line.

**What actually happens in this diff**
I introduce single-character-swap misspelling detection in a very specific case: that of 'class', 'trait', and 'interface' following modifiers 'abstract' or 'final'. I chose this case because it sidestepped most of the obstacles-to-general-implementation-of-spellcheck that I detail below.

**Obstacles to general implementation of spellcheck**
//First:// the current implementation of spell check relies on current_token_text, which returns the wrong text for tokens that have a leading newline. (For example, it'll return "\ncla" for token 'class' that begins a new line, when it should return "class.") So, spellcheck implementation is blocked on tokens-with-leading-newlines until that issue is remedied. drphil tagging you because you mentioned you were working on location reporting in the FFP.

//Second:// Properly implementing spellcheck in the wide-ranging way I initially thought was possible--over huge match statements like parse_statement or parse_declaration--actually requires handling lots of different cases. In order to properly process a misspelling, the following things need to happen:
* Detect the misspelling
* Call skip_and_log_misspelled_token (meaning, push the misspelling as extra_token_error trivia, and add an error to the parser)
* If the target token is consumed at this stage, report it as missing (by "target token," I mean the anticipated token. So, if you received 'ckass', the target token is probably 'class'.)
* If the target token isn't consumed at this stage, send the parser down the appropriate series of function calls. Then, when the target token *is* eventually consumed, make sure it's reported as missing.

Note that in the case I choose to implement spellcheck on here, the target token is imminently consumed, which makes things easier. I also avoid the issue with current_token_text and newlines.

Unfortunately my internship is ending before I'm able to give enough thought to this problem to solve it :'(

Also, for contrast with the new errors I added, this is what errors used to be output on the definitions of C, I, and T:

  (3,10)-(3,10) The 'class' keyword is expected here.
  (3,16)-(3,17) Encountered unexpected token 'C'.
  (6,7)-(6,7) The 'class' keyword is expected here.
  (6,17)-(6,18) Encountered unexpected token 'I'.
  (9,16)-(9,16) The 'class' keyword is expected here.
  (9,22)-(9,23) Encountered unexpected token 'T'.

Differential Revision: D5645005

fbshipit-source-id: 5111dab3420de337f0fa4a545f748d9c0f9061f2