Static calls and late-bound-static calls
Summary:
This also includes explicit-implementation non-static method calls,
which still get parsed as static calls.

This requires us to have support for classrefs, so that's included as well. I
allocate classref ids in a similar way to local ids, with the restriction that
all classref slots must be dead at block boundaries. This means I can use a
dead-simple algorithm to allocate the IDs instead of having to do anything
sophisticated.

Reviewed By: mofarrell

Differential Revision: D5649477

fbshipit-source-id: 7f173c33dcce729b86f5040617f15ef46f817d04