Fix read from dead ActRec in fcallHelperThunk
Summary:
When fcallHelperThunk determines that the callee has been intercepted, and the
interceptor indicates to not run the callee, it retrieves the return address out
of the callee's ActRec and uses it to return to the caller. However, the ActRec
is dead at this point, since the interception handler will tear down the
callee's frame before returning. In fact, part of the dead ActRec now contains
the return value. Nonetheless, we get away with it most of the time, since
nothing overwrites the dead ActRec.

However, there is at least one situation where this doesn't work. Suppose one of
the arguments to the callee has a destructor, or the callee itself has a
destructor. We'll call the interceptor, determine we shouldn't call the
function, then tear down the callee's frame. At this point, the arguments to the
callee, as well as the callee's this pointer (if any) is in Variants inside
RunInterceptHandler. Since the callee's frame is torn down, the stack pointer
now points at the return value provided by the interceptor. Once we return from
RunInterceptHandler, the Variants go out of scope, possibly running a
destructor. These destructors will push a frame onto the stack at the current
stack pointer, overwriting what was left of the callee's ActRec. Once the
destructor returns, and we return to fcallHelperThunk, we'll retrieve a now
garbage return address and crash.

Fix this by doing the proper thing and not accessing the callee's ActRec once it
is dead. Instead, save the return address before calling the interceptor, and
return it back to fcallHelperThunk directly. To facilitate this, add a new
DestMode to vasm to allow for functions which return two values (returning a
struct in C++), and use it for fcallHelper. To further enforce that nobody
accesses this dead ActRec again, trash fill it in debug mode.

Reviewed By: mxw

Differential Revision: D5657323

fbshipit-source-id: 5d538aab9d12e40361efb820cc7760a845b8e770