Change ExternCompiler protocol to report a compiler version
Summary:
Right now we don't have a way for HHVM to know if an external compiler
has updated and we need to invalidate any cached units that were compiled using
the old version.

First, I added a step to the protocol between HHVM and the compilers that makes
the compiler report a version string on startup.

HHVM uses this version number when it mangles the unit hash, so pairing HHVM
with a different set of external compilers will result in those units being
recompiled.

This isn't really ideal since it's very conservative--e.g. changing the PHP7
compiler version when working with an all-Hack codebase would invalidate the
whole cache. That said, this is straightforward and correct and probably won't cause
problems in realistic environments where we're using one compiler or the other.

I haven't entirely worked out how the compilers will know their version string.
For the PHP7 emitter, I can write a script much like the current repo schema
generation in HHVM itself. That may work for hackc too. I invite input on other
ways to accomplish this. I also looked at GNU ld's `--build-id` but it appears
not to be portable at all.

Reviewed By: mofarrell

Differential Revision: D5455905

fbshipit-source-id: 9d9ea8cdc73055d0bdaca26033ef817fb21af61d