Pessimize `resolve_class` in HHBBC Index
Summary: In non-repo mode, even if there is only one `cinfo` in the index for a specific `php::Class`, we can not conclude that this `cinfo` is the resolution, since we haven't seen other `cinfo` that migh pop-up later. This diff pessimize our resolution by returning a name-only `res::Class`.

Reviewed By: mxw

Differential Revision: D5586658

fbshipit-source-id: 170e32b779ac5d9e3ca85ca4fe8892da267787fb