Set current function while emitting special class functions
Summary:
The emitter wasn't setting the current function before emitting the
"special" class functions (pinit, sinit, cinit, etc). For the most part, this
was harmless, but one side effect of this is that the first opcode in the
function isn't considered reachable. This is because one of the things the
reachability check considers is if the current bytecode is at the current
function begin offset.

Therefore, the logic to handle something like a ternary operation with a
constant condition wouldn't actually emit any bytcodes at all if its the first
expression in a function. The emitter would optimize away the condition (because
its a constant), but then fail to optimize away the true or false case (because
its considered unreachable due to the above). The emitter would then assert
eventually because we'd fail to push anything onto the stack where we should
have.

This can be most clearly with a class containing a single constant whose value
is a ternary expression with a constant condition.

Fix all of this by making sure that the emitter's current function state is set
properly before emitting these functions.

Reviewed By: intellectuallyswole, markw65

Differential Revision: D5665778

fbshipit-source-id: 8798a12686cfb633c8f00a9dd0a05535dbdf471d